Clean up constant cost stuff
